Seiya

+++ yoshida/holiday_app/holiday/views/result.py
@@ -8,11 +8,14 @@ def result():
 
 @app.route("/result/new_or_update", methods=["GET", "POST"])
 def new_or_update():
-    if len(request.form['text']) == 0:
-        flash('新規登録更新失敗：祝日・テキストが未入力です')
+    if len(request.form['date']) == 0:
+        flash(' 新規登録更新失敗：祝日・日付が未入力です。カレンダーから日付を入力してください')
+        return redirect(url_for('input'))
+    elif len(request.form['text']) == 0:
+        flash('新規登録更新失敗：祝日・テキストが未入力です。1~20文字で入力してください')
         return redirect(url_for('input'))
     elif len(request.form['text']) > 20:
-        flash('新規登録更新失敗：祝日・テキストは1~20文字で入力してください')
+        flash('新規登録更新失敗：祝日・テキストが20文字を超えています。1~20文字で入力してください')
         return redirect(url_for('input'))
 
     holi = Holiday.query.filter_by(holi_date=request.form['date']).first()
@@ -31,6 +34,9 @@ def new_or_update():
 
 @app.route("/result/delete", methods=["GET", "POST"])
 def delete():
+    if len(request.form['date']) == 0:
+        flash('削除失敗：祝日・日付が未入力です。カレンダーから日付を入力してください')
+        return redirect(url_for('input'))
     holi = Holiday.query.filter_by(holi_date=request.form['date']).first()
     if holi != None:
         db.session.delete(holi)
